+++ src/resources/drawable/map_pin.png
